feat: handle unit and application settings types
The unit and application settings for a relation is map[string]string,
but the params allows for map[string]any. Ensure that we correctly
convert them.

Also ensure that we can correctly handle a nil settings map. We don't
want to send a non-nil map back, as that means something else - wipe
the settings.