feat: add remote relation watcher on consumer model
This patch adds a new watcher on the cross model relation domain that is
going to be used in the firewaller to watch changes on the remote
relations in the consuming model.

The watcher watches all relations and filters out the ones that don't
have an application_remote_offerer associated to it.