fix(uniter): add support for peer relations
This patch fixes #20590 which prevents peer relation from being created. This was because
the Uniter API doesn't return the "otherApplicationName" in case of peer relation, which causes
various side effects, especially in `relation-get` and `relation-list` hook commands.

This fix simply returns the current application name as otherApplicationName for peer relation,
since this type of relation is one self.

- Updated `Relation` in `apiserver/facades/agent/uniter/uniter.go` to handle scenarios where the
  `otherAppName` for peer relations defaults to the current application name when only one endpoint
  is present.
- Added a new test, `TestPeerRelation`, in `apiserver/facades/agent/uniter/uniter_test.go` to
  verify correct handling of peer relations. The test validates that the relation details, life, and
  endpoints are accurately processed and returned.