feat: add custom trigger for relation units by relation endpoint uuid
There already is a trigger for relation_unit table based on the uuid. A
trigger based on the relation_endpoint_uuid of the relation_unit is
necessary for WatchRelationUnits.

It triggers on a insert or delete based on a specific
relation_endpoint_uuid. Update isn't required as the only update is to
departing, which isn't tracked in the watcher.