feat: implement WatchRelationUnits
It notifies if the application settings change, any unit relation
settings change, a relation unit is added or removed.

Relation units and the application settings can be watched based on the
relation endpoint uuid. However relation unit settings cannot. The
trigger for the relation unit settings is based on the relation endpoint
uuid. Each time the relation units are triggered, get all of the
relation unit uuids for the relation endpoing uuid. This list can be
checked when a relation unit setting changes to determine whether to
notify.