refactor: get unit k8s information service call
This commit refactors the service call for getting a units kubernetes
information. This was done so that we can correctly return the right
information to the caller and avoid needless bugs in the future.

While here I updated the return struct to not use network provider id.
The provider id value for a unit is not a network value in this context
and the implication should not be made.

The service now relies on the ordering returned by state. It would be
better if this wasn't the case but it will do for now.