refactor: storage domain error consts together
This places all the storage domain error consts together. There was
previously some attempt at trying to group errors types but the reality
is most of the error types break the grouping in one way or another.

The better way to group them is to document their purpose effectively so
a reasonable person can make the determination.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>