fix: child logger for aws s3 client
The AWS s3 client gets a juju logger and because the library has some
small format bugs in log message it makes the Juju uniter code look like
it has an error when in reality it is coming from aws.

This change creates a child logger for the aws s3 client so we can see
easily from logs that is not a direct Juju source of the problem.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>