feat: add remote relation watcher on offering model
This patch adds a new watcher `WatchOffererRelations` that will watch
(remote) relations on the offering side of CMRs. This watcher is going
to be used in the firewaller, which will trigger the changes for the
offering model branch.

This watcher is a stateful one. We chose this implementation to be more
efficient. The reason is that we don't want to hit the db on every
change in the relation table to be able to filter the non-remote
relations. This state (or cache) allows us to be more efficient, but we
need to watch also the application_remote_consumer namespace.