fix: revoke unit leadership only upon successful deletion
The unit removal job was running up unitl the deletion point immediately
when the unit was dying, but then returned to be rescheduled.

This caused leadership to be lost too early for the unit to action its
dying phase.

Now we only do it after the unit deletion is successful.