refactor: storage provisioner watcher tests for single transaction
We have a dead lock condition in watcher tests where if multiple changes
are made that trigger events these events may not end up in the same
term causing the the harness step to stall.

To fix this issue each harness step makes sure that the events we care
about are all done in one transaction.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>