feat: use same relation UUID on offering and consuming sides of CMR
This patch removes the application_remote_relation table, and makes sure
that the created synthetic relation on the offering side has the same
UUID as the one on the consuming side.
This is possible because when registering, the consuming side passes its
relation UUID.

This changes are needed to facilitate how the firewaller publishes
network changes from the consuming to the offering side: we now don't
need to wait for the offering side to notify when it's ready.