feat: implement & wire up remote relation removal
Removing a 'remote relation' (i.e. a relation between an app and a remote
app) is very similar to that of regular relations.

However, there are some small optimisations/changes that are needed for
this this case. This is because we create synthetic units for the
relation data to flow between and such. These units do not have a uniter
in this model, so cleanup needs to be handled here.

When a remote relation is removed, set the synth units to dead
immediately. When removing the remote relation, delete these units. We
do not need to schedule jobs or anything.

This required a few changes/fixes to the relation domain, to keep us
safe from calling the wrong method.

- RelationExists will now return false (& an error) if the relation is a
  remote relation.
- UnitsInScope will not return synthetic units. The justification for
  this is that the units are not 'meaningfully' in scope in this model.