feat: implement WatchRelationChanges watcher
The domain watcher must be wrapped no matter what to translate the
domain structure into a params version. Therefore use an existing
notifiy watcher and wrap it to get the changes needed and send them
along. This removes the need for a second domain watcher.

The second return value from WatchRelationChanges,
params.RemoteRelationChangeEvent, will be filled in once a PR
implementing PublishRelationChanges lands. The same data is returned.