refactor: rename storage not found error
Renames the storage not found error to storage instance not found. The
reason for this is storage is a transparent term in the model and
doesn't identify any single thing.

Storage is a class that the charm offers. Storage instance is the
instantiation of that class which is what we care about.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>