feat: add method to check if a relation is cross model
This commit adds a domain method to check if a given relation is cross
model or not.
This is done by checking if one of the applications in the relation is
found in either the application_remote_offerer or
application_remote_consumer tables. Thus, this method can be used on
both sides of the CMR consistently.