fix(provider/openstack): handle subnet availability zones
This patch fixes #20956, which occurs because in openstack, sometimes networks are not shipped with
availability zones. This causes an issue on bootstrap since a machine is provisioned with an
expected availability zone that should have been inserted in DQLITE before the provisioning.
The IAASAgentFinalizer call machineService.SetMachineCloudInstance with
bootstrapMachineHardwareCharacteristics which contains an availability zone. If this availability
zone hasn't been populated, the bootstrap fails.

It happens when networks are not linked to the availability zone, because the openstack doesn't have
any requirement one it. In this case, the openstack provider wasn't populating any AZ in database,
which triggered the bug.

This patch ensures that a subnet will be inserted with at least an availability zone. It is done
that way because there is no other codepath to insert AZs. So, if the associated network doesn't
have any AZs, the subnet is linked with all AZs known by the project, which solve the issue.

- Added `TestSubnets_PopulatesAZsFromNetwork` and `TestSubnets_PopulatesAZsFromNovaFallback` in
 `internal/provider/openstack/networking_test.go` to verify subnet AZ assignments.
- Modified `makeSubnetInfo` in `internal/provider/openstack/networking.go` to accept `azNames` for
  AZ propagation.
- Updated `Subnets` in `internal/provider/openstack/networking.go` to gather available AZs from
  Nova and handle scenarios where subnets lack associated AZs.
- Enhanced mock in `internal/provider/openstack/network_mock_test.go` for `ListAvailabilityZones`.
- Extended `NetworkingNova` interface in `internal/provider/openstack/networking_interfaces.go` for
  AZ listing support.