feat: implement state func for transforming from storage ids to uuids
The service layer for storage requires the ability to translate from
storage ids to attachments. This is commonly used from the facade layer
where it works with tags.

This state layer implementation allows the caller to get a storage uuid
that can be used in the rest of the controller.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>