fix(domain/modelconfig): remove space check for set model config
This patch removes the space check when applying the model config to a model.
This fix an issue that prevent creating new model when bootstraping with a `--model-default` file
containing a space on MAAS.

The root cause is that the space doesn't exists when applying this checker since spaces are created
by the provider.