refactor: rename relation watchers
The watchers names only told half the story, as in, you needed to know
that you where dealing with a relation in the first place. We've long
said that we're not doing the repository pattern, and that the name
of the entity needs to be represented in the name i.e. using
GetRelation instead of Get. Thus the watcher method name should include
the word relation.