fix(charm-cleanup): move resource and charm cleanup outside of removal transaction
Before the cleanup for dangling resources and charms was done in the same removal transaction for
unit/application. This is a problem because if we have any constraint error while trying to remove
charm (and we do) it will rollback unit/application removal. Which is not what we want. Now the
resource and charm cleanup is done in another separate transaction and if it fails we will log a
warning.