refactor: rename AddRemoteApplicationConsumerArgs
To AddConsumedRelationArgs. This better reflects what is happening as
the application and relation are added.

Disambiguate the offer and consumer applications and endpoints.

There can only be one endpoint for the consuming application has the
data represents an already created relation. Change to a single Relation
rather than a slice.