feat: add the ability to get a storage instances attachments
This adds the ability to the storage service for getting all of the
attachments that a storage instance has. i.e all the places the storage
instance is attached to a unit.

In the current Juju model this will most likely always be one result.
But we do have the concept and technically support shared storage so it
is plausible that down the line there could be more than 1.

This new service func is setup to support the multiple case making it a
richer interface long term for callers. It also forces callers to deal
with this fact now.

This work was done in support of the storage detach facade func. This
facade func allows for the caller to not supply the unit id thus
detaching the storage from all units and not just a specific case.