feat: wire in a logger to the storage facade
This commit wires in a logger for the storage facade. This decision was
made to support detach storage work as based on the composition of
arguments supplied by the caller different operations have to be
performed.

The logger will help debug and also verify behaviour when other
introspection options are not available.