feat: support detaching storage from all units
When doing the initial implementation of detach storae in the facade it
was assumed that the caller was always supplying a storage id and a unit
for which to detach storage from.

However not all clients do this. The juju cli client for example will
only supply the storage id and leave the unit empty. Where the python
client will supply the unit id to remove the storage off of.

This creates essentially two code paths for the facade. When no unit has
been supplied the expectation is that the storage will be removed from
all units that it is attached to. As of this writing the most likely
case is that the storage will only ever been attached to one unit.
However nothing in our modeling prevents the storage from being attached
to multiple units. In fact storage itself in the charm has a flag to
indicate that storage can shared amoung units.

The second code path is where the caller does supply a unit for the
detach. In this case the action will be to only ever remove the storage
from that unit leaving in place any other attachments the storage may
have.