feat(removal): introduce tombstone status for volumes and fs
Volumes and Filesystems have a strange relationship to life.
They have unfinished business after the are dead, such that the
storageprovisioner worker needs to possibly obliterate them.

After the storageprovisioner is done handling the filesystem or
volume's removal, it marks the storage entity as in the tombstone
status. At this point, in a non-force removal job, the removal
domain can delete the tombstone'd filesystem or volume.