feat: uuid for synth applications
This is a massive change, esp. as such late notice. Essentially, we
need to have unique synth applications ids, that are the same as offer
connection uuids, but the application tokens are saved in the remote
application consumer under consumer_application_uuid.

This is a painful change, but it allows us to have multiple relations
to the same syth application.