feat: use GetRelationKeyByUUID for relation key
Relation keys should never be manually created when the data will be
saved or go over the wire. Ordering is based on the relation roles, thus
should come from the Key type to be consistent always.

Incoming relation keys may not be consistent and thus the parts of used
to find the relation