docs: attempt to update detach-storage docs to clear up wording
I have updated the cli command docs for detach-storage to remove the
unit/application reference as that is not the value that is supplied to the
command.

I also made it clear that the operation detaches storage from all units that it
is attached to. This lends it self to supporting shared storage down the line.