refactor: machine volume provisioning param types
The information required from state and that which is returned by the
service have divergence. This now requires the service layer to do a
fair amount of computation for the caller. To represent this information
new internal structs have been introduced to get the information out of
state so the service layer can correctly compute the result for the
caller.