refactor: charm storage DDL attributes to not be null
This refactors the shared attribute of charm storage to not be bull.
When we have boolean values it makes no sense to introduce a third state
for them with NULL as the meaning is ambiguous and leads to sql that
doesn't work when the value is NULL.

Easier to default these values with sqlair inserts.