feat: service logic implementation for machine provisioned volume params
This commit implements the service layer for getting machine provisioned
volume params out of the storage provisioner. This func only returns
volumes and attachments that model provisioned and have also not been
provisioned as of yet.

This logic behind this implementation was taken directly from 3.6 and
re-implemented for the dqlite world of Juju.