refactor: tidy up of filesystem attachment params
In an attempt to support multiple filesystem attachments for a unit of
the same storage name we need to re-work filesystem attachment params a
little.

The new structure now exposes the charms recommended location when
avalable. The mount point attribute now directly reflects the mount
point that has been set on the storage attachment if any.

I also renamed the sqlair struct used to better reflect the column
values being extracted as they were very ambigious with respect to the
DDL schema.