feat: calculate mount point for filesystem attachments in the model
This commit extends on the previous work done to provide filesystem
attachment params to the storage provisioner. In the first path we took
the storage location of the charm and used that as the mount point for
the filesystem attachment.

This approach is not in line without how we did things in 3.6 and it
also suffers from two problems. The first is if the charm has no
location set for the storage we would have ended up supplying an empty
value for the attachment mount point. The second problem is when a unit
may have more than 1 attachment for the same charm storage. In this case
two or more attachments would me made to the same mount point ending up
in an error state on the host.

To deal with this we need to calculate an idempotent mount point for
each attachment. The mount point is calculate on the fly with attachment
params because it is first fed into the storage provisioner which does
the mounting. When the mounting has been completed the provisioner
reports back to the controller the mount point which then becomes set on
the attachment.

This workflow also informs Juju that the attachment has been infact
provisioned if the question has to be asked. Attachment mounts points
are calculated in an idempotent manner using the unique uuid of the
attachment as a suffix.

The last part of this is if the charm has not specified a recommend
location for the mount point then the default storage location for Juju
will be used as the prefix in the path.

The implementation we had for 3.6 would also produce an error if the
charm had specified a location that fell within the default prefix used
by Juju /var/lib/juju/storage. I did not re-implement this check as it
mostly existed not gaurantee conflict with the calculation. Now that we
are using uuids this is very unlikely in will never matter.