refactor: calculate kubernetes filesystem mount points in template
This commit changes how the caas application worker has to think about
mount points for filesystems. Now when the service supplies back the
filesystem template it has pre-calculated all of the mounts points for
the worker.

This means the worker now contains less business logic about how to make
a mount point and it is just following the prescribed representation
from the template.

This further allows for the removal of the mount point calculation code
off of the storage domain.

Because data returned from state and service have diverged this commit
also introduces a new internal type to represent what the state knows so
that it can be expanded upon by the service layer.