feat: make caas application provisioner use filetemplate for mounts
The filesystem template recieved from the storage provisioner domain now
includes the mount points for the template that the caas application
worker can consume. This allows the worker to not have to do the mount
point computation itself with the logic being driven by the domain.

There is more wins here we can have long term with the template
providing more of the attachment information.