feat: add method to set status to error in relation
This commit adds a new method on the relation domain to update the
relation status to error. This is needed for the firewaller to set the
relation to error when the ingress network changes cannot be published
to the offering side.
This new method is only allowed on cross-model relations.