refactor: include storage name in machine volume provisioning info
It was miseed that the storage name is required in the machine volume
provisioning params because it is needed to calculate the volume tag for
the facade.

While here I noticed that I was also not returning the storage id
correctly which is also required to form the correct tags for the
volume.