test: fix issue with generation of id values for storage entities
Storage entities being created in tests were using the wrong value for
storage ids. When these tests were written it was assumed that this
would follow the tag value where in reality it is just the numerical
sequnce.

The volume, filesystem and storage instance helper funcs now also set
the correct value based off of the sequence table.