test: refactor storage instance creation helper
The storage instance creation helper was only returning the storage uuid
of the storage instance. We now have tests that are going to need the
storage id as part of the contract under test.

This change updates the helper and calls to now consider the storage id
that is returned.