refactor: adds storage name volume uuid to machine vol attachment
It is required that the storage name be passed up with machine volume
attachment provisioning params. This is so the volume tag can be
correctly constructed.

We also require the volume uuid to be passed up so that comparison
checks between volumes and attachments can be correctly made and not
done on volume id.