refactor: updates the machine storage provisioning volume params
Change is to include the new changes to params where the volume uuid and
storage name are now included in the output. While here I noticed that
we were incorrectly making the storage instance tag. The storage
instance tag now resembles the value that would have been returned from
the names package.