refactor: machine service hardware characteristics and availability zone
This work came out of moving the machine provisioner facade over to
internal errors. It was noticed that there was a call to hardware
characteristics for the availability zone.

The code was treating the not provisioned error as a machine not found
error. When in actual fact a not provisioned error was never possible
because we always populate the table for new machines.

All of the values for hardware characteristics are optional and it is
really there lack of existence that tells us not povisioned. However the
caller really didn't care about this and what it needed was the
availability zone not found error.

I moved the call over to the availability zone getter that already
existed.

I made the get hardware characteristics always return a struct and not a
pointer. Returning the pointer was a poor implementation hack from mongo
that doesn't need to exist in reality.

I also made the two functions check for the machine exists and validate
the uuid. All of the associated tests were testing that we were broken
in this area.