feat(removal): cascade offer removal
Offers are entities subordinate to applications. An offer 'offers' an
endpoint for an application.

So to remove an application, we need to remove all it's offers. Refactor
the removal domain for applications to remove these offers.

This required a small change to the cmr domain. To protect against
races, we need to ensure that we cannot establish new connections offers
of dying applications.