fix: initial changes missing for relation network egress
This commit fixes the relation network egress watcher which had
partially correct initial events. It was missing from the model config
and unit addresses sources.

Also, it renames the method `IsApplicationConsumer` into
`IsApplicationLocal`, which makes more sense since we're only looking at
the charm source to identify if the application is non-synthetic. This
makes the firewaller code more readable where it's used.

Another fix was done on the multiple workers started for the remote
relations in the firewaller: the watcher IDs were improved and added to
the structs that hold the workers, before they were simply the relation
tag.

Finally, some logs were added for future debugging to the firewaller.