fix: tweak deploy revision refresh test
The refresh revision test was failing because the test was running `juju refresh` and checking for the string "Added" to indicate the command suceeded.
In Juju 4 the output of refresh when the app is still deploying is,
```
$ juju refresh juju-qa-test
Added charm-hub charm "juju-qa-test", revision 21 in channel latest/edge, to the model
ERROR setting application "juju-qa-test" charm: unit "juju-qa-test/0" not in space "alpha"
```
The correct thing to do is wait for the app to be ready then run the refresh.