fix: avoid empty spaces from distorting output
In Juju 3.6 network addresses with empty spaces are filtered out.
In Juju 4 we were not doing that and on AWS a network interface with 2 addresses has the public address in an empty space while the private address is in the "alpha" space. Joining the 2 together resulted in the string "alpha " (note the extra whitespace at the end).
As an extra, I've removed the `spaces` field from the `statusContext` struct since it was no longer populated or used.