feat: notify machine when storage is deleted from machine during agent dying
When a machine agent is dying, it tries to EnsureDead on itself, if an
error is returned, the machine needs to retry. Currently machines only
retry when the machine life watcher notifies. This change enhances the
existing watcher which included its own life and the life of units on
the machine. This now notifies on the removal of a machine_filesystem,
machine_volume or a storage_filesystem_attachment,
storage_volume_attachment or storage_volume_attachment_plan that shares
this machines net_node_uuid.