fix: flaky TestRemoteRelationProviderRoleConsumingSide test in firewaller
This commit fixes a flaky test seen failing in CI although very
difficult to reproduce locally.
The fix consists of adding a sync point for waiting the rules to be
flushed after the machine is added. This is consistent to what is being
done in other tests in the suite (like `TestNotExposedApplication` and
`TestExposedApplication`).