fix(domain/schema): machine unit life trigger never record unit deletion
This patch fixes the behavior of machine unit life trigger when a unit is deleted.

Before this patch, it never generates any event while a unit is deleted, since it fetch the unit
table AFTER the deletion.

After this patch, it uses the net_node_uuid of the deleted unit to get the right machine name, which
 is the right behavior.

This fix a possible race while destroying a model with --force: In this case, unit can be deleted
when dying, so the machiner never get triggered for their dead.
Since the machiner doesn't put as dead a machine with undead unit, we could end in a state where the
 machine never get destroyed.

As a fly-by, I also update all the trigger code, removing the Join between machine and unit table
since we can just use the net_node_uuid from both unit and machine table.