refactor: support singleton filesystems on kubernetes
For filesystem attachment mount points in Kubernetes we tried to to
remove single storage attachments and always have an attachment prefix.
This works well but it broke the postgres charm because they hard code
the path in the charm.

The charm should be getting the mount point from the controller.