feat: pass server host for macaroon discharge
Originally we didn't want to hardcode the server host into the macaroon
itself. This can cause issues if the server IP address changes. At first
we wanted to use the controller UUID as a way to do this, but we can't
as it's not a real host name. We would require a DNS setup for this
to work nicely. We're just not there yet! For now we jump through all
sorts of hoops to get this working!

If the discharge handler fails, we'll have to some how cache the
cross model auth context per host and see if we can reuse the one
we've already got.