refactor: filesystem attachment params from state
This commit refactors filesystem attachment params from state to address
some concerns on the service layer. I have added the charm max count
value from the database so the service layer knows the maximum number of
storage attachments that are possible. This is required to establish if
the charm storage is a singleton or not.

I also renamed the ReadOnly attribute on the struct to reflect that this
value comes from the charm and not the filesystem attachment. The charm
dictates what it would like where as the value on the attachment is what
happens.

It is import to distinguish between where the values come from so logic
mistakes are not based on the wrong modelling.