refactor: service layer for filesystem attachment params
Re-works the storage filesystem attachment mount point logic to now
consider if an attachment is a singleton or not. This was done to help
keep storage attachments working for charms that exist already.

Specifically we see that the postgresql charm does not obey Juju storage
and hard codes the mount path. Because of this Juju is unable to change
the attachment information without breaking the charm. The charm should
be asking Juju for the information but it is not.

This makes used of the new parameters being passed up from state to
establish if a storage instances is a singleton.