fix: kubernetes to be driven off of attachments from domain
This commit fixes a bug we had in 4.0 where volume attachments were not
being made on to the right directory for the charm container. This was
because we had split brain logic for mount path calculation. This change
now makes sure that attachments and their mount points are calculated at
the domain layer and Kubernetes is just following the plan layed down.