fix(removal): cross-provisioned volume backed filesystem fast removal
When a filesystem is volume backed, and the volume is provisioned by
the model, it is impossible for the filesystem to go to dead, since,
a machine does not own it and will never run the finalisation of the
filesystem giving it the tombstone status.

This bypasses the tombstone status when there is no provisioner to
de-provision the filesystem.