feat: new charm storage validation for application domain
This commit adds a new validator to the application domain for
validating the storage definitions of a charm. The current validation
performed is making sure that a charm has not requested a filesystem
location that would violate Juju's default storage location.

Previously this check was done well into storage provisioning when
calculating the mount points of a filesystem attachment. This is not
ideal as it allows the controller to take on bad data that cannot be
achieved. It also provides a poor user expeirence for the end user as
the error will only surfact through the controller logs and with no
status indication.

By now doing the check up front at deploy time the user can recieve
better feedback.