fix(application-channel): set application-channel primary key as application-uuid
Before the primary key was spanning all fields in the table, this was causing issues because when we
update `application_channel` during `SetApplicationCharm` (which is called by `juju refresh`) we
can't set the `ON CONFLICT` clause just on application_uuid. This is needed to make sure we update
the channel for an application and we have a single row in application_channel for each application.