feat: handle missing life
If the life is missing or is an empty string from the API call, assume
that the life is alive. This is because the 3.6 code doesn't guarantee
that the life is always sent over the params. In this instance we don't
want to treat the absence of life as death.