fix(tests/model): update the ubuntu base version in `metrics.sh`
This patch changes the base Ubuntu version from 22.04 to 20.04 in `tests/suites/model/metrics.sh`
for the deployments of `app-one`, `juju-qa-test`, and `juju-qa-dummy-subordinate`.

There are issues at the moment with 22.04 snap which prevent ubuntu from starting and making the CI
fail.
It is not possible to use a 24.04 base or not defining the base because of subordinate relation
which requires compatible bases, and dummy-subordinate charm didn't have been released for noble.