fix(tests/model): update the ubuntu base version in `metrics.sh`
This patch changes the base Ubuntu version from 22.04 to 24.04 in `tests/suites/model/metrics.sh`
for the deployments of `app-one`, `juju-qa-test`, and `juju-qa-dummy-subordinate`.

We need to specify an explicit base to avoid breakage if charms are not released at the same pace.

There are issues at the moment with 22.04 snap which prevent ubuntu from starting and making the CI
fail.