fix: uniter facade availability zone information
The uniter facade availability zone implementation when moving over to
Dqlite was not implemented correctly based on the conctracts that were
offered for 3.6.

The first was that if an availability zone was not available it would
return a not provisioned error. This is an impossible case because for
the unter to be running it means the machine has been provisioned.

It is not an error if the machine has not AZ set. In fact this is
expected because not all clouds support this information. In this case
we should just return 3.6.

The reason we have never seen this problem until now is because the
machine domain offered an availability zone not set error that would
never have been returned. It wasn't until the domain logic was fixed to
be correct did this error actually surface showing this problem.

I have implemented the facade back to what it should be and also allowed
for non machine units to call the facade getting and get an empty AZ
string back. What this will allow us to do is remove another case out of
the uniter where we check if the model is caas or not.

I also took the time here to implement unit tests that do something. The
unit test that existed were testing contracts that didn't exist and also
domain errors coming back out of the facade.