feat: architecture slice helpers for agent binary domain
This commit adds to the agent binary domain two new functions for
working with multiple architecture values. The first func allows
transforming a slice of agent binaries to a slice of architectures.

The second func helps perform a difference over to architecture slices
to get the values from a which do not appear in b.

This work was done to support controller upgrades.