test: adds missing tests for http authorizers
This commit adds missing tests for the http authorizers that we use to
authz requests. After CVE-2025-0928 we added a new set of authorizers to
the controller to mitigate the security issue. However tests were never
included with the fix despite being available at the time.

This commit adds the missing tests and also moves the authorizers to a
new file so they are more clearly defined for the reader.

I also deprecated the tag kind authorizer as it should never be used in
practice.