fix: periodically cleanup idle connections
The http.Transport object defaults to keeping all idle connections forever if
the IdleTimeout is 0 and/or the MaxIdleConnections is 0. The DefaultTransport
uses a value of 100 idle connections, and an IdleTimeout of 90s. We can be a
bit more aggresive about this, because we don't really reuse the Transport
object.
Arguably another good step would be to use a shared Transport across the same
Juju controller. We need a different Transport for a different controller,
because we don't expect their CA Cert to be the same, but we could otherwise
cache the Transport.
However, we can also update our controller side so that it forcefully resets
idle connections periodically.  There doesn't seems to be a native Go way to
tell it to cap idle connections at a given timeout (eg server-side set 2min per
idle connection), so instead we just reset all idle connections with a
controller-config, defaulting to 10 minutes.

Note that both the client side fix and the server side fix indepedently "fix"
the leaking file descriptors. However, doing both is sane, since it will
protect us from other clients and other servers (eg cross model relations with
a 3.4 controller).

https://bugs.launchpad.net/juju/+bug/1952942

Include changes to Read and Write timeouts. Once the websocket takes
over, the logic changes, but at least while a unit is connecting, don't
wait forever for read or write to complete.

Also fix up the test suite for the httpserver worker.