feat: adds a new controller model authorizer to the controller
This authorizer was made to support agent binary uploads to the
controller. We need the ability to special case the controller model so
that uploads to the controller model require super user permissions.

This is because uploading to the controller model will make available
agent binaries for the entire controller and has a very large security
foot print.

The authorizer is written in such a way that if the authorization
request is not for the controller model it will pass the request through
to a fall through provider.